common/types/配下にui-elementsとui-partsフォルダ作成
