+++ common/types/ui-elements/field.ts
@@ -1,0 +1,14 @@
+export type TextInputProps = {
+  name: string;
+  fullWidth: boolean;
+  multiline: boolean;
+  minRows: number;
+  required?: boolean;
+  requiredMessage?: string;
+  label: string;
+  placeholder: string;
+  type: string;
+  onChange?: React.ChangeEventHandler<HTMLInputElement>;
+  onClick?: React.MouseEventHandler<HTMLDivElement>;
+  value?: string | number;
+};
