+++ common/types/ui-parts/form.ts
@@ -1,0 +1,18 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export type FormType =
+  | 'updateUser'
+  | 'createStack'
+  | 'createStackIntrospection'
+  | 'showStackIntrospection'
+  | 'updateStackIntrospection'
+  | 'createTeam'
+  | 'inviteTeam'
+  | 'updateTeam'
+  | 'createOutput'
+  | 'createOutputComment';
+
+export type ResetFormValueProps = {
+  setFormOpen: Dispatch<SetStateAction<boolean>>;
+  setIsRegisterEvent: Dispatch<SetStateAction<boolean>>;
+};
